Redirect to invites after the candidate viewed the confirmation page (#11412)
Previously when a candidate first opted in to the pool they would see a confirmation page with a lot of details about the pool. And with every update they would see the page again.

This commit uses the notifications table to only redirect the candidate to the confirmation page once.