[lb/1095 support board] error processing course change decisions (#11406)
## Context

Support request
https://trello.com/c/QFxCaxZN/1095-error-processing-course-change-decisions

We are syncing duplicate codes sites from publish. It's a problem, but
largely unimportant to us because all the user sees is the site code and
they duplicates are weeded out when we create the course options (we
don't get sites for the provider, we get the sites for a particular
course, and skip over any duplicate site codes).

On the API, however, we don't restrict our search for valid site codes
to the given course. So a vendor will get a "multiple sites match that
code" error.

## Changes proposed in this pull request

When we look for the site code, only look for the selectable course
options relevant to the course, not for the provider.

## Guidance to review



## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency