Allow monthly statistics report to be run with different dates (#11428)
## Context

There have been pipeline issues and we believe that is why we haven't
been update to run the latest Monthly statistics report. We need to be
able to test using dates from last week, when the last pipeline ran.

## Changes proposed in this pull request

This pr allows us to use bespoke dates for generating the monthly
statistics report.

## Guidance to review

<!-- How could someone else check this work? Which parts do you want
more feedback on? -->


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency