Candidate with multiple application_forms can recover their account (#11435)
Previously, a small number of candidates, about 40 up till now. Have not been able to recover their candidate accounts since we introduced one login because of a foreign key constraint on the application_form.

When we recover someone's account, we destroy the old one but for 40 candidates that have more than 1 application form linked to their candidate record this wasn't possible.

Adding dependent: :destroy on previous_application_form will allow candidates like this to recover their accounts in the future.

In addition, I've added an extra step to our account recovery form, to make sure we don't allow any application_forms with submitted applications to de deleted. This check is already in the controllers but not in this form directly. It's also in a spec now.