Vendor API - Update documentation to reflect recent changes (#11403)
## Context

We made changes to the API but we did not document it with a version
change and release notes.

This caused a problem with at least one of our vendors (technology 1).

We explored the possibility of releasing a 1.7 with the new changes but
we have decided to draw attention to additions to 1.6 in the release
notes.

## Changes proposed in this pull request

Add recent API additions to the release notes and highlight them in the
reference documentation:

The changes are:
- Add PreviousTeacherTraining to the ApplicationAttribute
- Add optional course parameter to confirm-deferred-offer endpoint

## Guidance to review

Please visit the following URLs and review the text relating to new
changes (third draft text is pictured below):
-
https://apply-review-11403.test.teacherservices.cloud/api-docs/release-notes
-
https://apply-review-11403.test.teacherservices.cloud/api-docs/v1.6/reference#important

#### Image 1:
<img width="582" height="623" alt="Screenshot 2025-11-13 at 15 58 07"
src="https://github.com/user-attachments/assets/a32b103b-f465-481b-b81b-4a8fc0d4302d"
/>

#### Image 2:
<img width="843" height="650" alt="Screenshot 2025-11-18 at 11 33 05"
src="https://github.com/user-attachments/assets/bc313981-6bc8-4026-9bbf-6b1c640bf1cb"
/>

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency