Log session error if the backchannel endpoint returns 500
The one login team has flagged that they are getting some 500 errors
from us when hitting the backchannel endpoint in apply.

I've tested the endpoint and it works fine for my account.

It's quite hard to diagnose the issue without some debugging
information. So this commit adds some errors to the session_error table.

We add errors to this table to not push sensitive data to sentry.

I'll keep an eye on for these errors in the future.