Investigate possibility of announcing FAC search results (#11450)
## Context

When a user completes a search, there is nothing to announce the number
of results that have been returned. This is especially unhelpful when
there are 0 results and the user has to tab through several other page
elements to discover this.

## Changes proposed in this pull request

- Update page title with the search results.

This is an imperfect solution as the page title is not consistently
announced on page refresh following a search but if the user is aware
that the title updates they have the option to tab up to it rather than
tabbing down through multiple other page elements.

Other solutions attempted yielded similarly inconsistent results:
- Visually hidden text at the top of the page (including attempting to
add different roles, such as `"alert"`).
- A stimulus controller set to read out a visually hidden div on page
reload making use of `role="status"`, `aria-live="assertive"`,
`aria-atomic="false"`. This was slightly more reliable but it is not
best practice to rely on javascript in these instances.

## Guidance to review

- Visit the FAC index and run cmd + F5
- Observe the screen reader behaviour

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency