Remove feature flag for duplicate site codes (#11440)
## Context

Due to changes in the publish application, we now have duplicate site
codes for providers. We do not, however, create duplicate site codes on
a course level (ie, if the provider has two sites with code AA, the
first AA might be attached to one course, and the second AA might be
attached to another course, but both AAs will not be attached to a
single course).

This was causing an issue on the API when finding a course option based
on codes. It was fixed by scoping the site search to the course. Now
that it has been tested in sandbox and activated in production, we can
delete the feature flag.

## Changes proposed in this pull request

Remove uses of the feature flag. A future PR will remove it from the
database.

NOTE: we'll wait until monday afternoon to merge this, when we are
confident it doesn't need to be reverted.

## Guidance to review


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [x] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency