Remove gem update system from docker.
This is blocking our deploys. It tries do update rubygems to 4.0 and
there are gems incompatible with this version.

Updating our bundler to 4.0 doesn't fix this issue.

Don't really know why we need to call 'gem update --system' but it's not
very common to do this in docker files. So will remove it for now.