Fix nationalities form, in personal information (#11477)
We don't use the correct list of countries and domiciles in our personal information forms.

When candidates input their nationality, they can input values like Puerto Rico and Taiwan. Whilst they cannot input these values in their address, as their domicile.

The UK does not recognise these countries, but they are valid domiciles. So we need to stop candidates choosing a nationality like Puerto Rico and allowing them to choose this as a domicile.

This PR allows that. There are also, candidates in production with this data, that needs to be fixed after this PR. Luckly the number of candidates in this position is under 20.


-- Also removed the gem update --system step from dockerfile.
This is blocking our deploys. It tries do update rubygems to 4.0 and
there are gems incompatible with this version.

Updating our bundler to 4.0 doesn't fix this issue.

Don't really know why we need to call 'gem update --system' but it's not
very common to do this in docker files. So will remove it for now.