Look for a monthly report in the whole month before raising error
We raise a sentry error if the monthly statistic report is not generated
on the generated date. This is good but if we eventually generate this
report after the generated date, this error will still be raised.

This commit will raise a sentry error only after we checked if we
generated this report at any point within the month.