Lockdown omniauth-rails_csrf_protection gem to 1.0.2
This gem provides a mitigation against CVE-2015-9284

Currently if we update to 2.0.0 our specs will break. Looking at the
issue tracker for the gem, it seems like other people have the same
problem. There might be a compatibility issue with rack, another gem.

For now, we can just keep the v1 version of this gem. There are no major
changes to the gem in v2. The protection is still there in v1.