Lockdown omniauth-rails_csrf_protection gem to 1.0.2 (#11486)
This gem provides a mitigation against nvd.nist.gov/vuln/detail/CVE-2015-9284 (Cross-Site Request Forgery on the request phase when using OmniAuth gem with a Ruby on Rails application)

Currently if we update to 2.0.0 our specs will break. Looking at the
issue tracker for the gem, it seems like other people have the same
problem. There might be a compatibility issue with rack, another gem.

This gem also allows our system and request specs to be marked as secure requests by omniauth. Otherwise they would be raised as insecure.

For now, we can just keep the v1 version of this gem. There are no major
changes to the gem in v2. The protection is still there in v1.