Qualifications should use degree countries
We've recently fixed a bug that allowed candidates to select places that
the Foreign Office doesn't recognize as countries. Like Puerto Rico.

In fixing this bug, we introduced another one. Where they can select
these teritories as countries for qualifications.

This fixes this issue and renames the methods to be more descriptive.