Degree grade interruption - Part 1 (#11487)
## Context

Some candidates do not meet the entry requirements for the courses they
apply for.

We don’t want to prevent them from applying because some providers can
still accept candidates that don't have the required grade, but have
different work experience.

However, we do want to warn candidates about the mismatch before they
submit, to help them think about whether it is the right course to apply
for.

## Changes proposed in this pull request

- Add new interruption step to the review and submit application flow 
- Delete the component which previously rendered degree grade
information and warning and replace with `DegreeGradeEvaluator` service

nb. There is a second flow for this interruption page described on the
card, namely when the candidate enters a degree grade and has draft
application choices whose required grade is higher than the highest
degree grade, including the one has just been saved. I have decided to
address this in a separate PR to first get some feedback on this
implementation and for general reviewing purposes given it will carry
its own complexity and edge cases.

## Guidance to review

- Sign in as Deandre Kub:
https://apply-review-11487.test.teacherservices.cloud/support/applications/68
- Attempt to submit the draft application
- You should see two interruption pages, one is for the short personal
statement, the second is to say that her highest degree grade is lower
than that required for the course

- Now return to the candidate index
-  Add a Master's degree for Deandre
- Next time you begin the review and submit journey, you should no
longer see the degree grade interruption
- Delete the Master's degree

- Add a choice for a TDA 
- You should not see the degree grade interruption page


https://github.com/user-attachments/assets/a523002f-25d6-4757-9e7a-cc5cced3ffa3

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency