Remove fee details from 2026 courses (#11505)
## Context

In Publish / Find, fee_details were rolled over, but not actually
updated by providers. It is not visible on the find publish UI, but it
is still coming through on the API, which means we are showing
candidates 2025 data on 2026 courses.

The other team is aware and are going to make a change to the API and
rollover process, but in the meanwhile, let's clean up our data.

## Changes proposed in this pull request

An earlier PR will make sure we always set fee_details to nil when
syncing from the API. This PR is a migration to remove existing data.
There are about 7400 records to update.

## Guidance to review

<!-- How could someone else check this work? Which parts do you want
more feedback on? -->


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency