Degree grade interruption - Part 2 (#11506)
## Context

Some candidates do not meet the entry requirements for the courses they
apply for.

We don’t want to prevent them from applying because some providers can
still accept candidates that don't have the required grade, but have
different work experience.

However, we do want to warn candidates about the mismatch before they
submit, to help them think about whether it is the right course to apply
for.

See PR #11487 for the first interruption page within the application
review and submit flow.

## Changes proposed in this pull request

Add an interruption page to be triggered during the degrees wizard:
- When a candidate has a draft application choice and the highest degree
grade they have is lower than the required grade it will show an
interruption at the end of the degree wizard stating "Your degree grade
does not match the eligibility criteria for the course you have drafted
an application for" with some guidance
- If they have multiple application choices in draft, we compare the
highest application choice required grade with the highest candidate
grade and show an interruption stating "Your degree grade does not match
the eligibility criteria for one or more of the courses you have drafted
applications for" with some guidance
- The interruption will also appear when editing a degree if these
conditions are met
- Additionally, now that we have the interruption page for those to whom
it applies, we only need to show the hint text in the complete section
component if the grade data is not one of the structured options in the
form

The interruptions won't appear if there are no application choices in
draft; nor if the candidate has a Master's/Doctorate degree or an
undergraduate degree of the required level; nor if the course applied
for doesn't require a degree (or a particular grade).

The interruptions will only appear for countries where we hold
structured degree data (including UK and Irish degrees but also
countries where the degree structure corresponds - see the
`DegreeGradeEvaluator` for more on that logic)

## Guidance to review

- Log in as Latina Goodwin:
https://apply-review-11506.test.teacherservices.cloud/support/applications/70
- Add a completed bachelor's degree with a grade of Third-class honours
- Review the interruption page that comes up at the end of the flow
(just after entering the award year)
- You will see the content referring multiple choices
- This page will also appear whenever you edit any individual step in
the degree wizard from the review page

-- 

- Now delete one of the choices 
- Go back to 'Degree'
- Click the change link on the grade step within the degree table
- Click 'Save and continue'
- You should now see the content referring to a single choice 

-- 

- Add a Master's and check that the interruption does not appear
- Remove the Master's
- Change the undergraduate degree grade to predicted and check that the
interruption does not appear
- Change the undergraduate degree grade to a 2:2 or above and check that
the interruption does not appear
- Change it back to a Third
- Delete the final application choice
- Go back to 'Degree'
- Click the change link on the grade step within the degree table
- Click 'Save and continue'
- You should return to the Degree review page without any interruption


https://github.com/user-attachments/assets/6e52c167-82cd-44e7-b196-512fa376a690

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency