Fix 'other nationality' error bug (#11508)
## Context

While on QA, we discovered an ambiguous error message when an invalid
nationality is added for 'other nationality': "If you have an additional
nationality, select it from the list" (see thread link to the Trello
card).

We also discovered that some raw inputs to the form were not being
handled: when trying to save an invalid nationality it will continue
through the journey without error and without saving.

## The bug


https://github.com/user-attachments/assets/24c290d1-e630-4476-a1cc-14e10d4a9c88


## Changes proposed in this pull request

- Update error messages to say 'Select your nationality' when blank
- Handle invalid raw data with error messages to say 'Select your
nationality from the list' when invalid

## Guidance to review

- Log into the review app as a candidate
- Start the 'Personal details' journey
- Try to break the form with various invalid inputs and a mix of valid
and invalid inputs
- Check that the expected results save



https://github.com/user-attachments/assets/d4ce7240-6275-4705-bc78-bd851980a18f



## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency