Change start date for performance recruitment reports (#11512)
## Context

The weekly performance report begins being published to providers in the
mid-cycle. (week 27, which is Monday 30 March 2026).

The Policy research team has done some discovery work on Data-sharing.
One of the findings is that the providers would like to see the
performance report earlier than is currently scheduled. Monday 12
January. (week 16) for the 2025-26 recruitment cycle year.

## Changes proposed in this pull request

- Changed first cycle week for the report from 27 to 16 (12 Jan 2026
this year)
- Added specs for `RecruitmentPerformanceReportTimetable`

## Guidance to review

<!-- How could someone else check this work? Which parts do you want
more feedback on? -->


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency