Use DFE reference data for finding the nationality for the API (#11516)
## Context

We have started using the DFE reference data for the Nationality
drop-down list, but did not change the data used for mapping the data in
the API. This brings the API mapping in line with the dfe reference data

It is still a problem that the dfe reference data is not consistent --
the docs say that it lists FDCO countries, but there are about 17
countries on the list (including Hong Kong) that are not FDCO countries.
This is being raised with the team responsible for the data.

Initial queries in blazer show that there are maybe 120 applications in
2026 that have selected a nationality that would not be matched and
reported in the API. A migration will come in a second PR.

## Changes proposed in this pull request

Use DFE reference data for nationality mapping since it is being used
for selection.
Changes to the carry over script to filter out unmatchable nationalities
in the future.

## Guidance to review

Sign in as a user without submitted applications (draft is good)
Go to Personal details and change the nationality to Hong Konger or
Cantonese
Now view the application as a support user, and click on 'Vendor API' to
view the application as it will appear on the API.
The nationalities array should include "HK"


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency