Lb/1558 add provider reports to support interface (#11518)
## Context

The Policy research team has done some discovery work on Data-sharing.
One of the findings is that the providers would like to see the
performance report earlier than is currently scheduled. Monday 12
January. (week 16) for the 2025-26 recruitment cycle year.

We plan on testing the generation on 5 January.

Also, they are recommending changes to the report itself.

## Changes proposed in this pull request

Currently, the reports are only viewable from the provider interface. We
want to move them to the support interface to make it easier to design
and test as changes are made.

This has involved moving a lot of provider_interface components into a
shared space so that they are available to support without causing
namespacing issues.

## Guidance to review

In the review app
Sign in as support
Find the provider U80
Navigate to the 'Recruitment performance reports' tab
You should see the data and a note that it isn't available to providers
yet.

Then go to any other provider
Navigate to the reports tab
You won't see any data, but you'll see the explanation as to when data
will be available


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency